+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version 0.4a-dev
 
+## New functions:
+02/05/2024
+- *processing.convolve*
+
 ## Modified functions:
 04/04/2024
 - *Spectrum_2D.__init__*: reshape of the fid is now made with last dimension implicit